Move expanding map logic out of world_map
also put Cell into a new file to avoid import errors
