+++ aimmo-game/simulation/game_logic/__init__.py
@@ -1,4 +1,10 @@
 from .spawn_location_finder import SpawnLocationFinder
-from map_updaters import ScoreLocationUpdater, MapContext, PickupUpdater
+from map_updaters import ScoreLocationUpdater, MapContext, PickupUpdater, MapExpander
 
-__all__ = ['SpawnLocationFinder', 'ScoreLocationUpdater', 'MapContext', 'PickupUpdater']
+__all__ = [
+    'SpawnLocationFinder',
+    'ScoreLocationUpdater',
+    'MapContext',
+    'PickupUpdater',
+    'MapExpander'
+]
