+++ aimmo-game/simulation/game_logic/map_updaters.py
@@ -4,6 +4,8 @@ import random
 import math
 from logging import getLogger
 from simulation.pickups import ALL_PICKUPS
+from simulation.cell import Cell
+from simulation.location import Location
 
 LOGGER = getLogger(__name__)
 
@@ -46,3 +48,30 @@ class PickupUpdater(_MapUpdater):
             if random.random() < world_map.settings['PICKUP_SPAWN_CHANCE']:
                 LOGGER.info('Adding new pickup at %s', cell)
                 cell.pickup = random.choice(ALL_PICKUPS)(cell)
+
+
+class MapExpander(_MapUpdater):
+    def update(self, world_map, context):
+        LOGGER.info('Expanding map')
+        start_size = world_map.num_cells
+        target_num_cells = int(math.ceil(
+            context.num_avatars * world_map.settings['TARGET_NUM_CELLS_PER_AVATAR']
+        ))
+        num_cells_to_add = target_num_cells - world_map.num_cells
+        if num_cells_to_add > 0:
+            self._add_outer_layer(world_map)
+            assert world_map.num_cells > start_size
+
+    def _add_outer_layer(self, world_map):
+        self._add_vertical_layer(world_map, world_map.min_x() - 1)
+        self._add_vertical_layer(world_map, world_map.max_x() + 1)
+        self._add_horizontal_layer(world_map, world_map.min_y() - 1)
+        self._add_horizontal_layer(world_map, world_map.max_y() + 1)
+
+    def _add_vertical_layer(self, world_map, x):
+        for y in range(world_map.min_y(), world_map.max_y() + 1):
+            world_map.grid[Location(x, y)] = Cell(Location(x, y))
+
+    def _add_horizontal_layer(self, world_map, y):
+        for x in range(world_map.min_x(), world_map.max_x() + 1):
+            world_map.grid[Location(x, y)] = Cell(Location(x, y))
