+++ aimmo-game/tests/test_simulation/maps.py
@@ -3,7 +3,8 @@ from __future__ import absolute_import
 from collections import defaultdict
 
 from simulation.location import Location
-from simulation.world_map import Cell, WorldMap
+from simulation.cell import Cell
+from simulation.world_map import WorldMap
 from simulation.game_logic import SpawnLocationFinder
